refactor: deduplicate logic
